+++ packages/nimble-components/src/table/styles.ts
@@ -1,6 +1,15 @@
 import { css } from '@microsoft/fast-element';
 import { display } from '@microsoft/fast-foundation';
-import { bodyFont, bodyFontColor } from '../theme-provider/design-tokens';
+import { White } from '@ni/nimble-tokens/dist/styledictionary/js/tokens';
+import {
+    applicationBackgroundColor,
+    bodyFont,
+    bodyFontColor,
+    fillHoverColor
+} from '../theme-provider/design-tokens';
+import { Theme } from '../theme-provider/types';
+import { hexToRgbaCssColor } from '../utilities/style/colors';
+import { themeBehavior } from '../utilities/style/theme';
 
 export const styles = css`
     ${display('flex')}
@@ -22,9 +31,48 @@ export const styles = css`
     .header-row {
         display: flex;
         flex-direction: row;
+        background: ${applicationBackgroundColor};
+        position: relative;
     }
 
     .header {
         flex: 1;
     }
-`;
+
+    .row {
+        background: ${applicationBackgroundColor};
+        position: relative;
+    }
+
+    .row::before {
+        content: '';
+        width: 100%;
+        height: 100%;
+        position: absolute;
+    }
+
+    .row:hover::before {
+        background: ${fillHoverColor};
+    }
+`.withBehaviors(
+    themeBehavior(
+        Theme.color,
+        css`
+            .header-row::before {
+                content: '';
+                width: 100%;
+                height: 100%;
+                position: absolute;
+                background: ${fillHoverColor};
+            }
+
+            .row::before {
+                background: ${fillHoverColor};
+            }
+
+            .row:hover::before {
+                background: ${hexToRgbaCssColor(White, 0.15)};
+            }
+        `
+    )
+);